Add files via upload
This simulation have a very simple circuit of positive clippers that takes the Alternating Current as input and then clips the  positive side of Alternating Current. In this way we get the Current that has a signal with only one side. This clipping Process is done by the help of diode. The diode has the ability to pass the current only in one direct. Hence clipping of one side occurs. The result is then shown on oscilloscope. By changing the direction of diode, one can change the whole circuit into a negative clipper.